# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fruit & Dry Fruits Shop Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories, admin]
      summary: Create a category (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCategory
      tags: [categories, admin]
      summary: Update a category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories, admin]
      summary: Delete a category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── PRODUCTS ──────────────────────────────────────────────────────────
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{itemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /cart/apply-coupon:
    post:
      operationId: applyCoupon
      tags: [cart]
      summary: Apply coupon code to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponApply"
      responses:
        "200":
          description: Cart with discount applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "400":
          description: Invalid coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── WISHLIST ──────────────────────────────────────────────────────────
...
  /contact:
    post:
      operationId: sendContactMessage
      tags: [contact]
      summary: Send a contact message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactMessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── COUPONS ───────────────────────────────────────────────────────────
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons, admin]
      summary: List all coupons (admin)
      responses:
        "200":
          description: Coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [coupons, admin]
      summary: Create a coupon (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Created coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/{id}:
    delete:
      operationId: deleteCoupon
      tags: [coupons, admin]
      summary: Delete a coupon (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── ADMIN: DASHBOARD ─────────────────────────────────────────────────
...
  /admin/messages/{id}:
    delete:
      operationId: adminDeleteMessage
      tags: [admin, contact]
      summary: Delete a message (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── ADMIN: INVENTORY ─────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── AUTH ──
...
    Category:
      type: object
      required: [id, name, slug, imageUrl, productCount]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        productCount:
          type: integer
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
        description:
          type: string
        imageUrl:
          type: string
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string

    # ── PRODUCTS ──
...
    CartItem:
      type: object
      required: [id, productId, productName, productImage, price, discountPrice, quantity, weightOption, subtotal]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: ["string", "null"]
        price:
          type: number
        discountPrice:
          type: ["number", "null"]
        quantity:
          type: integer
        weightOption:
          type: ["string", "null"]
        subtotal:
          type: number
    Cart:
      type: object
      required: [items, subtotal, discount, deliveryCharge, total, couponCode]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        discount:
          type: number
        deliveryCharge:
          type: number
        total:
          type: number
        couponCode:
          type: ["string", "null"]
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        weightOption:
          type: string
    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
    CouponApply:
      type: object
      required: [code]
      properties:
        code:
          type: string

    # ── WISHLIST ──
...
    ContactMessageInput:
      type: object
      required: [name, email, subject, message]
      properties:
        name:
          type: string
        email:
          type: string

        subject:
          type: string
        message:
          type: string
...
    Coupon:
      type: object
      required: [id, code, discountType, discountValue, minOrderAmount, expiresAt, isActive]
      properties:
        id:
          type: integer
        code:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        minOrderAmount:
          type: number
        expiresAt:
          type: string
          format: date-time
        isActive:
          type: boolean
    CouponInput:
      type: object
      required: [code, discountType, discountValue, minOrderAmount, expiresAt]
      properties:
        code:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        minOrderAmount:
          type: number
        expiresAt:
          type: string
          format: date-time

    # ── DASHBOARD ──
...